[FIX]: Issue when using user defaults  fix #2
The description of the issue is :
"Hi, when I switch to a different language, the app switches to that language and the language is saved for the next time. If I switch back to the DeviceLanguage the app switches to the Device Language on the fly. But after restarting the app, the app is translated into the language translated which I selected before switching back to the DeviceLanguage."
Thanks @akandor for pointing it out and finding the solution